Coerce messy, but convenient format of all logical sorts to a psOpenSorts object.

Only available for logical open sorts. Should only be used for hand-entered open sorts, administered on paper. This format has limitations; see note.

as_psLogicalOpenSorts(logical_open_sorts, ...)

# S3 method for matrix
as_psLogicalOpenSorts(
  logical_open_sorts,
  descriptions_messy = NULL,
  keep_LETTERS = TRUE,
  ...
)

# S3 method for data.frame
as_psLogicalOpenSorts(
  logical_open_sorts,
  descriptions_messy = NULL,
  keep_LETTERS = TRUE,
  ...
)

Arguments

logical_open_sorts
a character matrix with rows as items, columns as participants and logical category assignments as character strings in cells. Categories are identified by a subset from LETTERS, same as in descriptions_messy. Assignments must be the same subset of LETTERS as the column names in descriptions_messy. Rows and columns must be named.

For example, if some participant assigned her (self-described) categories A, D and Z to some item, the cell for that item and participant would read "A, D, Z". Order and punctuation are ignored.

...
further arguments passed to or from other methods.

descriptions_messy
a character matrix with rows as category indices, columns as participants and category descriptions in cells. Rows must be named by a subset of LETTERS to conveniently enter, and identify them from logical_open_sort. The row names are arbitrary identifiers, but will be retained for the canonical form. Columns must be named as participants.

Defaults to NULL, in which case no descriptions are available.

Notice category description in one row have nothing in common other than their indices: For example, the category descriptions in a row named 'B' are all by different participants, and may refer to entirely different aspects. They are only conveniently entered in a table, and all share the fact that they were the second description provided.

When some category has not been defined by the participant, the value in the cell should be NA. Empty strings "" will also be considered NA.

keep_LETTERS
a logical flag. Defaults to TRUE, in which case the LETTERS for the category descriptions and assignments are retained as names, even though they are just indices and not actual meaningful names (useful for debugging).

Details

The canonical representation of open sorts in psOpenSorts() can be cumbersome to enter manually. For logical (nominally-scaled) open sorts, a simpler, but messier format can be conveniently entered as two separate spreadsheets of descriptions_messy and logical_open_sorts using as_psLogicalOpenSorts().

Methods (by class)

matrix: coerce matrix or data.frame of all logical sorts to psLogicalOpenSorts

data.frame: coerce messy, but convenient data.frame or matrix with all sorts to psLogicalOpenSorts

Note

When a category is assigned, but never described, it is TRUE in the respective logical matrix entries and their description is NA: This is still considered valuable, if incomplete information. When a category is described, but never assigned, it is omitted from the data entirely.

When no category was assigned to some item in logical_open_sorts, an empty character string "" should be in the respective cell.

An NA value implies that the given participant never considered the given items at all, across all her categories. Notice this implies limited scenarios of NA for data entered in this messy, convenient form. The more complicated cases, where a participant did consider some, but not all items in the assignment of a category, or -- equivalently -- all categories in their assessment of all items, cannot be recorded in this convenience format. Such more granular NA records can, however, be recorded in the canonical data representation, where the respective cell of the items x category logical matrix would be NA. If your data gathering procedure produces such granular NA records, do not use this convenience function.