Validate S3 classes from this package.

Use check(), test(), assert() and expect() to validate classed objects from this package.

Check S3 class.

check(x)

test(x)

# S3 method for default
test(x)

expect(x, info = NULL, label = NULL)

# S3 method for default
expect(x, info = NULL, label = paste(class(x),
  "S3 class"))

assert(x, collection = NULL, var.name = NULL)

# S3 method for default
assert(x, collection = NULL, var.name = paste(class(x)[1],
  "S3 class"))

# S3 method for QItemStrata
check(x)

# S3 method for QPreSorts
check(x)

# S3 method for QPeopleFeatures
check(x)

Arguments

x
class object created by respective constructor function.

info
[character(1)] See expect_that

label
[character(1)] See expect_that

collection
[AssertCollection] If an AssertCollection is provided, the error message is stored in it. If NULL, an exception is raised if res is not TRUE.

var.name
[character(1)] The custom name for x as passed to any assert* function. Defaults to a heuristic name lookup.

See also

Other validation functions: QItemConcourse, QSorts, correlate, extract, make_grid, score

Examples

# create checkable object
x <- QItemConcourse(
  concourse = matrix(
    data = c(
      "Man lives to work.", "Man lebt, um zu arbeiten.",
      "Man works to live.", "Man arbeitet, um zu leben."
    ),
    nrow = 2, ncol = 2,
    dimnames = list(
      items = c("live_2_work", "work_2_live"),
      languages = c("english", "ngerman")  # ideally, these are valid babel languages
    )
  )
)

# check object and friends ...
check(x)  # returns TRUE or error message

#> [1] TRUE

test(x)  # returns TRUE or FALSE

#> [1] TRUE

assert(x)  # returns error or silently object