Create Co-Occurence Matrices.

Creates co-occurence matrices from logical q-category assignments.

count_cooccur(ass)

Arguments

ass
Named list of logical matrices, one for each participant. Each logical matrix has items as named rows, category indices as columns and logical values in cells.

Value

An integer array with items as rows and columns, participants as third dimension and cells as co-occurence counts.

Details

The diagonal is replaced with the maximum number of categories for that person, to standardize the entire table.

See also

Other import: import_qcat