Make pandoc tex variable option

This function creates pandoc variable key value pairs for LaTeX preamble.

declare_pandoc_var(key, value)

declare_pandoc_fontsize(fontsize_global = NULL)

declare_pandoc_linestretch(linestretch = NULL)

declare_pandoc_lang(lang = NULL)

declare_pandoc_geometry(
  paperwidth = NULL,
  paperheight = NULL,
  top = NULL,
  bottom = NULL,
  left = NULL,
  right = NULL,
  unit = "cm",
  vcentering = FALSE,
  hcentering = FALSE
)

Arguments

key
[character(1)] giving the key, such as "geometry".

value
[character(1)] giving the value, such as "margin=1in".

fontsize_global
[character(1)] giving the document-wide font size.

Must be one of:

NULL in which case the system default fontsize is used. (default)

'10pt'

'11pt' or

'12pt'

linestretch
[numeric()] giving the line spacing in multiples, e.g. 1.25, 1.5. Defaults to NULL for default LaTeX line spacing.

lang
[character(1)] giving a valid BCP 47 language code code, such as en_US.

Must be one of:

NULL in which case there is no multilingual support (default)

'ar-DZ' for arabic (Algeria)

'ar-IQ' for arabic (Iraq)

'ar-JO' for arabic (Jordan)

'ar-LB' for arabic (Lebanon)

'ar-LY' for arabic (Libya)

'ar-MA' for arabic (Morocco)

'ar-MR' for arabic (Mauritania)

'ar-PS' for arabic (Palestinian Territory)

'ar-SY' for arabic (Syria)

'ar-TN' for arabic (Tunisia)

'de-DE' for german

'de-AT' for german (Austria)

'de-CH' for german (Switzerland)

'dsb' for lower sorbian

'hsb' for upper sorbian

'el-polyton' for greek (polytonic)

'en-AU' for english (Australia)

'en-CA' for english (Canada)

'en-GB' for english (Great Britain)

'en-NZ' for english (New Zealand)

'en-UK' for english (United Kingdom)

'en-US' for english (United States)

'grc-ancient' for greek (ancient)

'la' for latin

'sl' for slovenian

'fr-CA' for french (Canada)

'pt-BR' for portoguese (Brazil)

'af' for afrikaans

'am' for amharic

'ar' for arabic

'as' for assamese

'ast' for asturian

'bg' for bulgarian

'bn' for bengali

'bo' for tibetan

'br' for breton

'ca' for catalan

'cy' for welsh

'cs' for czech

'cop' for coptic

'da' for danish

'dv' for divehi

'el' for greek

'en' for english

'eo' for esperanto

'es' for spanish

'et' for estonian

'eu' for basque

'fa' for farsi

'fr' for french

'fur' for friulan

'ga' for irish

'gd' for scottish

'gez' for ethiopic

'gl' for galician

'he' for hebrew

'hi' for hindi

'hr' for croatian

'hu' for magyar

'hy' for armenian

'ia' for interlingua

'id' for indonesian

'is' for icelandic

'it' for italian

'km' for khmer

'kmr' for kurmanji

'kn' for kannada

'ko' for korean

'lo' for lao

'lt' for lithuanian

'lv' for latvian

'ml' for malayalam

'mn' for mongolian

'mr' for marathi

'nb' for norsk

'nl' for dutch

'nn' for nynorsk

'no' for norsk

'nqo' for nko

'oc' for occitan

'pa' for panjabi

'pms' for piedmontese

'pt' for portoguese

'rm' for romanian

'ro' for russian

'sa' for sanskrit

'se' for samin

'sk' for slovak

'sq' for albanian

'sr' for serbian

'syr' for syriac

'ta' for tamil

'te' for telugu

'th' for thai

'ti' for ethiopic

'tk' for turkmen

'tr' for turkish

'uk' for ukrainian

'ur' for urdu or

'vi' for vietnamese

Used for multilingual typsetting support via LaTeX's babel package and others. Careful: Depending on the local tex distribution, not all valid languages may also be supported by LaTeX. Use check_latex_lang() to verify.

paperwidth, paperheight
[numeric(1)] giving the width and height of documents in unit. For good typographical results, should be as close as possible to the actual physical measurements of documents encountered by users. Defaults to NULL.

top, bottom, left, right
[numeric(1)] giving the margin in unit. Defaults to NULL.

unit
[character(1)] giving the units for the above dimensions.

Must be one of:

'cm' for metric (default) or

'in' for imperial

vcentering, hcentering
[logical(1)] indicating whether content should be vertically/horizontally centered. Defaults to FALSE.

Value

[character()] giving pandoc variable option, options in the case of geometry.

Functions

declare_pandoc_fontsize: declare base font size

declare_pandoc_linestretch: declare linestretch (to be passed on to LaTeX setspace package)

declare_pandoc_lang: declare language

declare_pandoc_geometry: declare options for LaTeX geometry package.