Infer maximum grid of q-sorting grid.

Infers the maximum q-sorting grid from observed dataset.

diag_distro_max(sorts)

Arguments

sorts
An integer array with item handles as first dimension, people as second dimension, arbitrary dimensions thereafter, and item positions in cells. Dimensions must be named.

Value

A positive integer vector of counts of length equal to cover the range of values, named by values.

See also

Other distribution helpers: diag_distros, diag_inside_grid, diag_same

Examples

diag_distro_max(sorts = civicon_2014$qData$sorts[,,"before"])

#> -7 -6 -5 -4 -3 -2 -1  0  1  2  3  4  5  6  7 
#>  1  1  2  4  6  9 10 11 10  9  6  4  2  1  1 

# true, because study used forced Q distribution