Forced distribution test.

Test whether all people-variables completely fill out the same grid.

diag_forced(sorts, grid)

Arguments

sorts
An integer array with item handles as first dimension, people as second dimension, arbitrary dimensions thereafter, and item positions in cells. Dimensions must be named.

grid
A positive integer vector of a length covering the range of values, specifying maximum allowed counts for each value. (in Q-parlance, the maximum column heights for the Q-sorts).

Value

A logical value of length 1, TRUE if the distribution is forced, else FALSE.

Details

If all people-variables display the same value counts, respondents plausibly faced a forced distribution, but this does not strictly speaking follow. It is unlikely, though conceivable, that respondents all adhered to precisely the same distribution, but could have used other available slots.

Whether, or not, a Q dataset should be considered "forced" can therefore only ascertained from a comparison with the physically available grid during the administration of the Q-sort.

In the above, improbable case, the function warns the user of possible misspecification.

Examples

sorts <- civicon_2014$qData$sorts[,,"before"]
diag_forced(sorts = sorts,
            grid = diag_distro_max(sorts))

#> [1] TRUE