Same distribution test.

Tests whether the value counts are the same for every people-variable.

diag_same(sorts)

Value

A logical vector of length one, TRUE when value counts are the same for all people-variables, else FALSE.

Note

Q data that displays the same value counts for all people-variables is likely to stem from a forced distribution, where respondents had to fill in all physically available slots in the grid, but this need not be so. Hypothetically, if unlikely, some people-variables might follow exactly the same distributions, even though they had more slots available. For more information, see diag_forced().

See also

Examples

diag_same(sorts = civicon_2014$qData$sorts[,,"before"])

#> [1] TRUE

# true, because study used forced Q distribution