Convert markup to vector formats

Call various tools with options to convert vector formats.

md2tex(path, fontsize_global = NULL, linestretch = NULL, lang = NULL, ...)

texi2pdf2(path)

pdf2svg(path, page = 1)

svg2grob(path)

md2tex_mem(
  x,
  path_in = "foo",
  fontsize_local = NULL,
  alignment = "justified",
  ...
)

texi2pdf2_mem(x, path_in = "foo", ...)

pdf2svg_mem(x, path_in = "foo", ...)

svg2grob_mem(x, path_in = "foo", ...)

Arguments

path
[character(1)] to an input file with or without extension

fontsize_global
[character(1)] giving the document-wide font size.

Must be one of:

NULL in which case the system default fontsize is used. (default)

'10pt'

'11pt' or

'12pt'

linestretch
[numeric()] giving the line spacing in multiples, e.g. 1.25, 1.5. Defaults to NULL for default LaTeX line spacing.

lang
[character(1)] giving a valid BCP 47 language code code, such as en_US.

Must be one of:

NULL in which case there is no multilingual support (default)

'ar-DZ' for arabic (Algeria)

'ar-IQ' for arabic (Iraq)

'ar-JO' for arabic (Jordan)

'ar-LB' for arabic (Lebanon)

'ar-LY' for arabic (Libya)

'ar-MA' for arabic (Morocco)

'ar-MR' for arabic (Mauritania)

'ar-PS' for arabic (Palestinian Territory)

'ar-SY' for arabic (Syria)

'ar-TN' for arabic (Tunisia)

'de-DE' for german

'de-AT' for german (Austria)

'de-CH' for german (Switzerland)

'dsb' for lower sorbian

'hsb' for upper sorbian

'el-polyton' for greek (polytonic)

'en-AU' for english (Australia)

'en-CA' for english (Canada)

'en-GB' for english (Great Britain)

'en-NZ' for english (New Zealand)

'en-UK' for english (United Kingdom)

'en-US' for english (United States)

'grc-ancient' for greek (ancient)

'la' for latin

'sl' for slovenian

'fr-CA' for french (Canada)

'pt-BR' for portoguese (Brazil)

'af' for afrikaans

'am' for amharic

'ar' for arabic

'as' for assamese

'ast' for asturian

'bg' for bulgarian

'bn' for bengali

'bo' for tibetan

'br' for breton

'ca' for catalan

'cy' for welsh

'cs' for czech

'cop' for coptic

'da' for danish

'dv' for divehi

'el' for greek

'en' for english

'eo' for esperanto

'es' for spanish

'et' for estonian

'eu' for basque

'fa' for farsi

'fr' for french

'fur' for friulan

'ga' for irish

'gd' for scottish

'gez' for ethiopic

'gl' for galician

'he' for hebrew

'hi' for hindi

'hr' for croatian

'hu' for magyar

'hy' for armenian

'ia' for interlingua

'id' for indonesian

'is' for icelandic

'it' for italian

'km' for khmer

'kmr' for kurmanji

'kn' for kannada

'ko' for korean

'lo' for lao

'lt' for lithuanian

'lv' for latvian

'ml' for malayalam

'mn' for mongolian

'mr' for marathi

'nb' for norsk

'nl' for dutch

'nn' for nynorsk

'no' for norsk

'nqo' for nko

'oc' for occitan

'pa' for panjabi

'pms' for piedmontese

'pt' for portoguese

'rm' for romanian

'ro' for russian

'sa' for sanskrit

'se' for samin

'sk' for slovak

'sq' for albanian

'sr' for serbian

'syr' for syriac

'ta' for tamil

'te' for telugu

'th' for thai

'ti' for ethiopic

'tk' for turkmen

'tr' for turkish

'uk' for ukrainian

'ur' for urdu or

'vi' for vietnamese

Used for multilingual typsetting support via LaTeX's babel package and others. Careful: Depending on the local tex distribution, not all valid languages may also be supported by LaTeX. Use check_latex_lang() to verify.

...
Arguments passed on to declare_pandoc_geometry

paperwidth

[numeric(1)] giving the width and height of documents in unit. For good typographical results, should be as close as possible to the actual physical measurements of documents encountered by users. Defaults to NULL.

paperheight

[numeric(1)] giving the width and height of documents in unit. For good typographical results, should be as close as possible to the actual physical measurements of documents encountered by users. Defaults to NULL.

top

[numeric(1)] giving the margin in unit. Defaults to NULL.

bottom

[numeric(1)] giving the margin in unit. Defaults to NULL.

left

[numeric(1)] giving the margin in unit. Defaults to NULL.

right

[numeric(1)] giving the margin in unit. Defaults to NULL.

unit

[character(1)] giving the units for the above dimensions.Must be one of:

'cm' for metric (default) or

'in' for imperial

vcentering

[logical(1)] indicating whether content should be vertically/horizontally centered. Defaults to FALSE.

hcentering

[logical(1)] indicating whether content should be vertically/horizontally centered. Defaults to FALSE.

page
[integer(1)] giving the page in the pdf to convert.

x
[character()] or [raw()] giving the input.

path_in
[character(1)] giving path to use for input file with or without extension. Defaults to "foo". Useful for debugging.

fontsize_local
[character(1)] giving a valid LaTeX font size.

Must be one of:

NULL in which case the default local fontsize is used (default)

'tiny'

'scriptsize'

'footnotesize'

'small'

'normalsize'

'large'

'Large'

'LARGE'

'huge' or

'Huge'

alignment
[character(1)] giving the alignment of the text.

Must be one of:

'justified' (default)

'left'

'right' or

'center'

Value

depending on the function postfix:

for the base functions, [character(1)] (invisibly) giving path to output file with extension. Exception: svg2grob() always returns a grid grob.

For _mem, [character()] or [raw()].

Details

Functions with the _mem postfix abstract away file system operations and have been cached via memoise::memoise().

Functions

md2tex: markdown to LaTeX via pandoc

texi2pdf2: latex to pdf via LaTeX

pdf2svg: PDF to SVG via pdf2svg

svg2grob: SVG to R graphics (grid) via grImport2::readPicture()

md2tex_mem: markdown to LaTeX via pandoc

texi2pdf2_mem: latex to pdf via LaTeX

pdf2svg_mem: PDF to SVG via pdf2svg

svg2grob_mem: SVG to R graphics (grid) via grImport2::readPicture()