Draw one random Q-sort dataset.

Draw one random Q-sort dataset, parametrized as original data.

draw_rand_sort(grid = NULL, n = NULL, p = NULL)

Arguments

grid
A positive integer vector of a length covering the range of values, specifying maximum allowed frequencies for each value. (in Q-parlance, the maximum column heights for the Q-sorts).

n
A positive integer vector of length 1, as the number of people-variables. Defaults to NULL, in which case parameter is inferred from data.

p
A positive integer vector of length 1, as the number of item-cases. Defaults to NULL, in which case parameter is inferred from data.

Value

A random integer matrix with item-cases as rows and people-variables as columns.

Details

This function draws a random dataset, parametrized as some real or hypothetical observed dataset from a Q study.

Q data has some peculiarities that must be implemented to draw comparable random datasets, and that are not addressed in existing packages implementing parallel analysis (such as paran):

Integers. The simulated data include only integers.

Draws without replacement. Because two items cannot occupy the same position in a Q-sort, random data must be drawn without replacement from the grid of available item positions (slots).

Forced or free distribution. Depending on whether the observed data resulted from a forced or free distribution, random data must be drawn from the appropriate available item positions (slots).

Helper function for run_parallel().

Note

Notice that in the case of a free distribution, there are no assumptions on the distributional shape (such as uni-modal, bimodal etc.). Random data for a free distribution are therefore drawn from an (on average over all people-variables) rectangular distribution over all available slots, as a neutral distributional shape. A more sophisticated parallel analysis may first infer some observed distribution parameters and then draw random data accordingly as per issue 146 on the development repo.

See also

Other parallel-analysis: run_parallel